feat: add generic segment group
