+++ src/main/java/com/jerryio/protocol_diagram/Main.java
@@ -120,7 +120,7 @@ public class Main {
         }
 
         if (args.print) {
-            System.out.println(diagram);
+            System.out.print(diagram);
         } else {
             doStartScanner();
         }
