+++ src/main/java/com/jerryio/protocol_diagram/diagram/render/Divider.java
@@ -1,32 +1,13 @@
 package com.jerryio.protocol_diagram.diagram.render;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-
 import com.jerryio.protocol_diagram.diagram.Field;
 import com.jerryio.protocol_diagram.diagram.render.element.DividerSegment;
 import com.jerryio.protocol_diagram.diagram.render.element.RowSegment;
 
-public class Divider {
-    private final int bit;
-    private final List<DividerSegment> segments;
-
-    private int used;
+public class Divider extends SegmentGroup<DividerSegment> {
 
     public Divider(int bit) {
-        this.bit = bit;
-        this.segments = new ArrayList<DividerSegment>();
-        this.used = 0;
-    }
-
-    public int getBit() {
-        return bit;
-    }
-
-    public Collection<DividerSegment> getSegments() {
-        return Collections.unmodifiableCollection(segments);
+        super(bit);
     }
 
     public void addSplice(RowSegment before, RowSegment after) {
@@ -38,9 +19,5 @@ public class Divider {
             segments.add(new DividerSegment(represent, used, length));
             used += length;
         }
-    }
-
-    public int getUsed() {
-        return used;
     }
 }
