+++ src/main/java/com/jerryio/protocol_diagram/diagram/render/Row.java
@@ -1,36 +1,13 @@
 package com.jerryio.protocol_diagram.diagram.render;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-
 import com.jerryio.protocol_diagram.diagram.Field;
 import com.jerryio.protocol_diagram.diagram.render.element.RowSegment;
 import com.jerryio.protocol_diagram.diagram.render.element.RowTail;
 
-public class Row {
-    private final int bit;
-    private final List<RowSegment> segments;
-
-    private int used;
+public class Row extends SegmentGroup<RowSegment> {
 
     public Row(int bit) {
-        this.bit = bit;
-        this.segments = new ArrayList<RowSegment>();
-        this.used = 0;
-    }
-
-    public int getBit() {
-        return bit;
-    }
-
-    public Collection<RowSegment> getSegments() {
-        return Collections.unmodifiableCollection(segments);
-    }
-
-    public RowSegment getSegment(int index) {
-        return segments.get(index);
+        super(bit);
     }
 
     public Row addField(Field field) {
@@ -49,23 +26,4 @@ public class Row {
             segments.add(new RowTail(used, bit - used, isVisible));
         used = bit;
     }
-
-    public int getUsed() {
-        return used;
-    }
-
-    public int getCount() {
-        return segments.size();
-    }
-
-    // public int[] getSplicePositions() {
-    //     int[] array = new int[segments.size()];
-    //     int index = 0;
-    //     for (int i = 0; i < segments.size(); i++) {
-    //         array[i] = index += segments.get(i).getLength();
-    //     }
-    //     return array;
-    // }
-
-
 }
